add a fledgeling meta build tool
I hacked that script together this afternoon and it is thus still far
from perfect: It should work if all modules to be build are in the
same source directory, though.

If I didn't miss anything important, the biggest priciple problem is
that out-of source builds cannot be supported well with the current
OPM build system because there is only the ${MODULE}_ROOT variable to
specify where a module is located, i.e., to get this job properly
done, one needs to be able to specify the directories on a finer level
(${MODULE}_INCLUDES, ${MODULE}_LIBRARIES, etc.). alternatively,
multiple ${MODULE}_ROOT directories could be allowed for the same module.

the current work around for this is sibling builds, which should be
removed ASAP in my opinion.