add support for bash command line parameter completion
as of now, this is limited to the eWoms tests, but once
OPM/opm-simulators#1512 has been merged, it will work for `flow` as
well without any change.

Since IMO parameter completion is a pretty big usability improvement,
we should probably add instructions how to enable it to the website
once OPM/opm-simulators#1512 has been merged.