2019-04.02 Bug Fixes
Bug fixes for when OPM Flow is not installed on the system and for when OPMRUN.ini is initially created by the program. Source code updated and new binary created

Fixed menu layout bug.
Fixed Project Directory bug for when the default OPMRUN.ini file is created.
Fixed Edit Parameters bug for when OPM Flow has not been installed.
Changed some test messages to be consistent with Options.