New Option: Generate Well Specification to generate WELSPECS, COMPDATA and COMPLUMP keywords.
The WELLSPEC option re-formats the WELSPECS and COMPDAT keywords exported from OPM ResInsight to be more  readable. At the same time, it writes out the OPM ResInsight Completion Data File, so that the data can be loaded back into to OPM  ResInsight. This is done so that in OPM ResInsight  one can declare one perforation for the whole well and run the export through this routine to get completions based on on reservoir unit. The module can optionally read an OPM ResInsight Formation Name Layer File and stores the date for determining the completion number in the COMPLUMP keyword. In this case COMPLUMP keyword is also generated.