Update opm_keyw.py to Handle Editing Files - WIP
Major re-factoring of code to implement loading of an input file and basic editing, still need to implement some features but the basic operations are working. For example, Ctrl+Z works as undo, but the right-click option does not.

Added a menu for various options plus a Generate menu item to generate a set of standard keywords for a section. Also changed how the various section filters are activated for easy of use.