Add Desktop Notifications to Background Jobs
Background jobs are run in a terminal process based on a script file generated by OPMRUN. Thus, there was no way to see the progress of the jobs as the runs progressed. Here we implement operating Desktop Notifications to inform the user of the progress. A job start a notification is issue with the OPM icon (front page of the OPM Flow manual). Upon completion a notification is sent based on the return code. If the return code is zero then the OPM icon is shaded green, if the return code is one then it is shaded red.

Works for both Linux and Windows 10.