Add Recursive Compression to OPMRUN Tools
Add recursive compression to compress all *.DATA jobs in a directory and below, as well as  ensemble recursive compression, that will compress all files in an ensemble directory. I needed this for a project that had far too many ensemble cases.

Added addition documentation and compression "help" option.

Fixed a bug with calling OPM ResInsight.

This was done some time ago I just needed to update the image in the manual and the web site for the commit.