Update package depdendecies
* PySimpleGUI -> FreeSimpleGUI
The package PySimpleGUI stopped providing open source support and changed its own license for a closed source. Later, the whole project was dismantled and not even closed source is possible. FreeSimpleGUI is a fork of the original open source PySimpleGUI with essentially the same possibilities.
The only real difference is that `file_type` argument on several functions expects a tuple of string pairs.

* pyDOE2 -> pyDOE3
Support for pyDOE2 seems to have stopped and is not compatible with current python versions. The fork pyDOE3 took over its development and seems to work exactly as before.